Ex 4.11. Print issue description to terminal editor
To print output to editor, we can write output to
a temp file then open the file with the editor.

Similar, to get input from an editor, we can create
a temp file, open the file with editor, then decode
the file to take necessary input, e.g. issue title
and body.