Ex 8.1. Clockwall
Modify clock2 to accept a port number, and
write a program, clockwall, that acts as a
client of several clock servers at once,
reading the times from each one and
displaying the results in a table, akin to
the wall of clocks seen in some business
offices. If you have access to geographically
distributed computers, run instances remotely;
otherwise run local instances on different ports
with fake time zones.