Ex 8.5. Benchmark CPU-bound task mandelbrot render
Exercise 8.5: Take an existing CPU-bound sequential program,
such as the Mandelbrot program of Section 3.3 or the 3-D surface
computation of Section 3.2, and execute its main loop in parallel
using channels for communication. How much faster does it run on
a multiprocessor machine? What is the optimal number of goroutines
to use?

Answer:
For non-heavy CPU-bound tasks, running in parallel is unlikely to
help. It might even be slower.