Ex 9.4. Pipelining goroutines
Exercise 9.4: Construct a pipeline that connects an
arbitrary number of goroutines with channels. What is
the maximum number of pipeline stages you can create
without running out of memory? How long does a value
take to transit the entire pipeline?

```
=== RUN   TestRunGoroutinePipeline
=== RUN   TestRunGoroutinePipeline/size_is_1
Time taken for size: 1 is 2.311µs
=== RUN   TestRunGoroutinePipeline/size_is_10
Time taken for size: 10 is 47.189µs
=== RUN   TestRunGoroutinePipeline/size_is_100
Time taken for size: 100 is 133.042µs
=== RUN   TestRunGoroutinePipeline/size_is_1000
Time taken for size: 1000 is 574.663µs
=== RUN   TestRunGoroutinePipeline/size_is_10_000
Time taken for size: 10000 is 3.488306ms
=== RUN   TestRunGoroutinePipeline/size_is_100_000
Time taken for size: 100000 is 32.504754ms
=== RUN   TestRunGoroutinePipeline/size_is_1_000_000
Time taken for size: 1000000 is 396.343255ms
=== RUN   TestRunGoroutinePipeline/size_is_10_000_000
Time taken for size: 10000000 is 3.822613301s
=== RUN   TestRunGoroutinePipeline/size_is_20_000_000
Time taken for size: 20000000 is 8.160054683s
=== RUN   TestRunGoroutinePipeline/size_is_30_000_000
Time taken for size: 30000000 is 12.865994869s
=== RUN   TestRunGoroutinePipeline/size_is_40_000_000
Time taken for size: 40000000 is 21.79272666s
=== RUN   TestRunGoroutinePipeline/size_is_50_000_000
Time taken for size: 50000000 is 2m40.284881758s
=== RUN   TestRunGoroutinePipeline/size_is_60_000_000
^Cgo tool test2json: signal: interrupt

Process finished with the exit code 1
```