Ex 10.4. Report packages in the workspace depend on give packages
Exercise 10.4: Construct a tool that reports the set of all
packages in the workspace that transitively depend on the packages
specified by the arguments. Hint: you will need to run go list
twice, once for the initial packages and once for all packages.
You may want to parse its JSON output using the encoding/json
package