Sec 11.6. Example functions: Add ExampleIntSet_Add
Other than TestXxx and BenchmarkXxx, ExampleXxx is also
another type of function in test. It aims to give a sense
of how to use a function; when coming with godoc, it provides
an interactive code on godoc. If there is `//Output:`
comment, the Example function becomes test. 'go test'
tests Example functions and asserts the actual output
with expected output in the comment.